requirements.txt
torch
torchvision
dlib
scikit-image
easydict
PyYAML
dominate
dill
tensorboardX
scipy
opencv-python
einops
PySimpleGUI

numpy
Pillow
matplotlib

ninja
torch
torchvision
opencv-python
numpy
scikit-image
scipy
pillow
tqdm
cog

fastai==1.0.51
wandb
tensorboardX==1.6
ffmpeg-python
youtube-dl>=2019.4.17
jupyterlab
opencv-python>=3.3.0.10
pillow>=8.0.0
